server/requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
djangorestframework-simplejwt
social-auth-app-django
openai
requests 